Merged PR 24: feat: center canvas in viewport
The motivation for this was that since we created a landing page and made the canvas smaller it feels kind of odd to click the anchor on the landing page then only scroll down half a viewport and still have half of the landing page in view.  So, to remedy this, the canvas section now takes up the entire viewport.
feat: center canvas in viewport

![image.png](https://dev.azure.com/CS360/c418b9c3-afc7-4f09-b157-e113e5ff9c0e/_apis/git/repositories/0821758c-7a5b-496e-aa7d-e9133f476984/pullRequests/24/attachments/image.png)

Related work items: #340